refactor(cockpit): extract proof evidence model
